Add tests for monotonic inputs with limit errors
This commit adds tests of limit expressions in top-k evaluation over
monotonic inputs. The top-k variants tested can be either monotonic or
basic, but for consistency with the input type, we add these tests to
top-k-monotonic.td. The commit also opportunistically makes a small
change to a comment in rendering.