sql: refactor WMR CTE constraint, key error checks
We checked that the derived and proposed RelationTypes in WMR CTEs
had compatible definitions for NOT NULL columns and keys. However,
the proposed column types cannot contain the NOT NULL constraint,
nor can we express keys.

Convert the potentially confusing errors we returned previously
to internal errors that should surface in testing when these
features get added.