persist-txn: rename LD flag and cli flag to persist_txn_tables
The original name was enable_persist_txn_tables, but a previous PR
switched this from a bool to an enum. I thought that would be safe,
since it hadn't been used, but now all sorts of things in CI are brake
if I try to turn it on because previous images expect it to be a bool.

Also, "enabled_" usually indicates a bool flag in mz, so this new name
probably avoids some confusion.