Tweak to consolidation of persist_sink correction
This commit introduces a small change to the consolidation policy of
the correction buffer in the write_batches operator of persist_sink.
The previous consolidation moments are extended to also include
doublings of the correction buffer. This should not be of much
effectiveness when inputs are consolidated, except for timing issues
with batch readiness, but can provide benefits when inputs are not. At
the same time, the overhead introduce is expected to be manageable in
the scenarios where the buffer grows, especially for hydration.