catalog: report progress sources as running
Progress subsources only reported as 'created' because they are
not hooked up to any form of health operator.

Instead of synthesizing the 'created' status for them, we could
instead synthesize the 'running' status to make their statuses
align with other sources.

Coincidentally, this is the same approach taken for webhook
sources, so there is some prior art in this space.