catalog: Update name and comment of PersistHandle (#23732)
This commit updates the name of PersistHandle and the surrounding
comments to more accurately convey what it is used for. It used to be a
wrapper struct around both a read and write handle, hence the name
PersistHandle. Now it is used exclusively for pre-open bootstrap catalog
reading and opening a catalog, hence the name
UnopenedPersistCatalogState.