storage: introduce ceasing status
If a source encounters a definite error, we currently expect them
to never run again.

However, the previous set of status had no way of encoding this
expectation, which meant that we would continually reset the
status to starting/running on restart.

To encode the permanence of this status, introduce the "ceased"
status and make a best effort at ensuring that the status is
never overwritten by any other statuses.