sql: lazily fill in PG source timeline ID during purification
It is possible that a PG source does not have a timeline ID because
the migration which added it to existing PG sources can fail. If
we detect that the migration failed, fill it in the next time we
are altering the source's publication details (i.e. during the next
ALTER SOURCE).