catalog: Implement data migrations for persist  (#23442)
This commit implements a data migrations framework for the persist
catalog. Like the stash version, this framework allows us to update the
on disk serialized representation of catalog data in the persist
catalog. While both the stash and persist catalog implementations
exist, we will have to write separate migrations for each
implementation. Additionally, we had to delete all previous proto
versions that did not contain the StateUpdateKind message.

Resolves #22977