sql: split Kafka source and sink options
Splitting the parsing and planning of Kafka sources and sinks is a net
simplification for the code. Even though a few options are shared
between sources and sinks (TOPIC, GROUP ID PREFIX), most are not. Having
separate types for sources and sinks avoids a separate validation step
to determine whether each option is valid for the context.

This is a pure refactoring commit.