coord: use one shared PostgresTimestampOracleConfig
Before, we were re-creating it from its components when needed.

This is a step towards dynamically changeable pg timestamp oracle
settings. The shared config, which holds a shared `DynamicConfig` makes
it so all created timestamp oracles share the same config, and we can
update that one when system vars change.

We do not yet add the code that actually does update the
`DynamicConfig`.