persist-txn: ensure each unique batch's serialization is unique
The code to handle retracting applied batches assumes that the encoded
representation of each is unique (it works by retracting and cancelling
out the raw bytes). It's possible to make that code handle any diff
value but the complexity isn't worth it.

So ensure that every committed batch has a unique serialization.
Technically, I'm pretty sure that they're naturally unique but the
justification is long, subtle, and brittle. Instead, just slap a random
uuid on it.