catalog: Add warning for specifying wrong schema
Materialize currently has a bug that all types are put into the
`pg_catalog` schema. For example, the following query works:
`SELECT 1::pg_catalog.mz_timestamp;`, but the following does not:
`SELECT 1::mz_catalog.mz_timestamp;`. `mz_timestamp` is in the
`mz_catalog` schema and therefore the second query should work and the
first should fail.

We can fix this bug and write a migration to fix the types in all
persisted views, but that will not let us know who in production is
manually specifying the `pg_catalog` schema for these types in
`SELECT` queries.

This commit adds a warn log to identify any time a user specifies the
`pg_catalog` schema for a type that should live in a different schema.

Works towards fixing #23833