transform: refactor optimizer notices
In order to persist notices in a `BuiltinTable` we need to be able
to emit retractions for a previously inserted notice (for example
when we delete the associated catalog item). However, the notice
structure of the notice vector emitted as a side-effect of the
optimization pipeline is not suitable for emitting retractions -
for example:

1. It doesn't have an optional `created_at` field.
2. Some of the text fields need to be resolved against catalog items
   that might be renamed / removed in the meantime.

In order to protect ourselves against invalid retraction issues
arising from these problems we make a distinction between the notice
type produced by the optimizer (`RawOptimizerNotice`) and the notice
type stored in the `CatalogPlans` struct (`OptimizerNotice`).

Also, since we expect to have more than two notice types in the
long term, we use the occasion to reorganize the code a bit so adding a
new notice is easier in the future. The required steps are as follows:

1. Create a `mz_transform::notice::<notice_type>` submodule.
2. Define a struct for the new notice in that module.
3. Implement `OptimizerNoticeApi` for that struct.
4. Re-export the notice type in `mz_transform::notice`.
5. Add the notice type to the `raw_optimizer_notices` macro which
   generates the `RawOptimizerNotice` enum and other boilerplate code.