adapter: Move mz_aclitem to mz_catalog
This commit updates the schema of mz_aclitem from mz_internal to
mz_catalog. Due to
https://github.com/MaterializeInc/materialize/issues/23833, mz_aclitem
has actually always been in the pg_catalog schema, and will remain in
the pg_catalog schema after this change. However, the type has
pretended to be in the mz_internal schema for long enough that it is
now stable, so it's ready to pretend to be in the mz_catalog schema.