catalog: make the persist impl's schema a Row with a single jsonb column
Technically not a Row, but rather a SourceData with no Errs, like we do
for tables.

This sets us up to expose the raw catalog shard as a queryable source in
mz SQL. This in turn would allow us to model something like
`mz_databases` as a `VIEW` (or `MATERIALIZED VIEW`) instead of manually
denormalize-ing catalog changes into a second place when committing DDL
changes.

There are a number of tweaks we could make to the proto->json mapping to
make it easier to actually defining these views in SQL, but the catalog
migration framework means we can "easily" punt that for a later time.

For now, start by modeling the top-level "kind" enum as [internally
tagged] to try to set persist pushdown statistics up for success. We
might later want to use [untagged] for all the proto oneofs, to make the
json more ergonomic, but this would require tweaking the protos a bit
(e.g. replace the GlobalId proto or its innards with the canonical
GlobalId string representation).

[internally tagged]: https://serde.rs/enum-representations.html#internally-tagged
[untagged]: https://serde.rs/enum-representations.html#untagged

Closes #23871