dbt-materialize: ensure get_empty_subquery_sql is valid SQL
Follow up to cb172dfa1, which in turn was a follow up to #23712. The
get_empty_subquery_sql macro is technically public API that users can
call directly, so we need to ensure that our hack to be able to split
the header from the body doesn't result in invalid SQL if the macro is
called directly.