compute: fix memory leak in delay logging
This commit fixes a memory leak in compute delay logging caused by the
logging dataflow relying on the assumption that import frontiers report
for a dataflow are always in advance of export frontiers reported for
the same dataflow so far. This isn't always correct because dataflow
operators can advance their output frontiers beyond their input
frontiers, and `persist_sink` in particular is known to do this.

The wrong assumption caused us to accumulate entries in the `time_deque`
when a multi-worker replica had materialized views installed.

This is fixed now by checking the export frontier first before pushing
a new entry to `time_deque`. If the export frontier is advanced past the
frontier we would push, we refrain from pushing it instead.