persist-txn: apply shard registrations (#23899)
Previously, after registering a data shard into a persist transaction
handle, the physical upper of the data shard was not moved. However,
the dataflow operator would wait for something something TODO(jkosh44)
make this commit message better. Then the dataflow operator is stuck
waiting for the physical upper of the data shard to advance to the
registration timestamp.

This commit adds the concept of applying a register, which will advance
the upper of the data shard to the register timestamp.

Touches #22173
Touches #22801