adapter: log redacted SQL if stuck on execute command
We had an incident today in which the coordinator got stuck due to a
pathological SQL query. We were lucky that the customer was able to
provide us with the query that caused the hang. In the future, we might
not be so lucky.

So, this commit teaches the coordinator watchdog to log the redacted SQL
that caused it to get stuck, if its stuck on a SQL execution command.
Careful introduction of an `Arc` here means the coordinator is net
neutral on deep clones of the SQL statement during command execution, so
there should be no observable performance impact to this change.

Fix #23938.