persist-txn: Apply shard forget
Previously, as part of executing a forget, we would physically advance
the upper of the forgotten data shard to the forget timestamp. This
commit removes this step from forget and introduces the concept
applying a forget, which physically advances the upper of a forgotten
data shard to the forget timestamp during transaction application.
We prevent the transaction shard from compacting past any unapplied
forgets so that the application is idempotent.

This approach is more in-line with how we treat registers and commits.

Touches #22173
Touches #22801