sql: store ServerVar default as Owned instead of &'static
Based on a comment, this appears have been added originally to avoid
some cloning. But it's somewhat restrictive of how you got the default
value and the currently avoided cloning doesn't happen enough in
practice to move the needle.

I've got an upcoming PR that reduces LD boilerplate and it isn't able to
produce a &'static ref. I was hoping to use that to motivate this
commit, but I spent 1/4 of my skunkworks today just rebasing this
commit, so pulling it out early. Plus, it looks like Joe also wanted to
do this at some point (#20367), so it's not just me :).