*: update various feature flag defaults
This commit adjusts some feature flags by setting their defaults to the
production defaults they have in LD. This makes it more likely that
development and testing happens on a configuration similar to
production, and that we catch bugs related to features enabled in
production.

It's possible that some of these feature flags should be removed too.