notice: emit redacted optimzier notice fields only if needed
A lot of times the redacted `message`, `hint`, and `action` fields of a
notice won't differ from the non-redacted counterparts. In those cases,
set the `redacted_~` field to `NULL` and use `coalesce(redacted_~, ~)`
in the `mz_notices_redacted` view in order to pick up the right value.

This should reduce the amount of memory required to keep optimizer
notices in both `CatalogPlans` and in the `mz_~notices~` catalog items
by ~50%.