persist-txn: Make apply idempotent on closed shard
Applying a transaction is meant to be idempotent, i.e. applying a
transaction multiple times should succeed and not affect the
correctness of the system. Previously, if someone tried to apply a
transaction on a data shard after closing that data shard, then the
system would panic. This commit fixes the described issue by returning
successfully from application on a closed shard.

We know that if a data shard is closed, then all transactions must have
already been applied, so it's safe to return immediately.

Fixes #23964