doc/developer: refine feature lifecycle
Split the checklists for progressing to Public Preview and General
Availability into two to avoid circularity; the requirements *for*
moving are different than what you need to do to facilitate the move.

Also adjust the wording of "basic {observability|metrics}" with
"suitable {observability|metrics}", to admit the possibility of a
feature being too small to warrant observability/metrics.