reclock: fix read hold leak leading to large in-memory traces
When a reclock handle was cloned its current read hold was (correctly)
re-added to the mutable antichain tracking the minimum compaction
frontier. When a reclock handle was dropped it first compacted itself to
the empty frontier in order to release its requirements on the frontier.
This would be correct if only for the fact that
`ReclockFollower::compact` only actually compacts if the handle is
initialized. This means that if a handle was shared and dropped before
the handle had a chance to initialize then we would start accumulate
reclock bindings in-memory, which make every future reclocking operation
continously slower. This PR fixes this issue and adds a unit test.

This issue looks related to #22128 but I'm not convinced yet that they
are the same because the only way I've found to trigger this bug is to
use a source that causes a drop of a `ReclockFollower`. The sources that
have this behavior are the ones that ignore the feedback mechanism to
commit offsets back to the upstream system but both Kafka and Postgres
sources actively use that.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>