webhook: Cache "webhook appender" from Coordinator  (#23777)
This PR refactors our webhook HTTP handler to cache the necessary info
for appending to a webhook source.

Previously for every webhook request we would send a message to the
Coordinator and get back the necessary metadata for appending to a
webhook, i.e. a channel to push updates and an expression to run
validation. Webhook sources rarely change (if ever) so sending a message
to the Coordinator on _every_ append is unnecessary, and effectively
serialized all webhook requests. In this PR we add a cache to the HTTP
server where for a given `(database, schema, name)` tuple, we store a
channel for appending and the expression for validation. The channel in
the cache gets closed if the underlying webhook source ever changes, at
which point we call into the Coordinator to get a new channel.

This removes the Coordinator from the critical path for webhook appends,
which reduces CPU load on `environmentd` and increases our throughput
for webhooks. Benchmarking with the same setup as
https://github.com/MaterializeInc/materialize/pull/21766, we are able to
sustain ~2,000 writes per-second at ~30% CPU usage, with spikes to 80%.

Note: I validated this approach with @mjibson to make sure we stay
aligned with the Platform V2 vision.
Note2: I didn't do any profiling, but the CPU spikes seem to follow a
similar shape as memory spikes, I have a branch stacked on top of this
one which should improve this.

### Race Conditions

Given we have multiple async components running concurrently, operations
can get interleaved. I believe we retain correctness though because
before any `ALTER` command would return, e.g. `ALTER SOURCE ... RENAME
TO ...` we invalidate the `MonotonicAppender`s. And the only way for
`Adapter` clients to obtain new `MonotonicAppender`s, is to go through
the Coordinator.

For example, the following scenario is impossible:

1. Run `ALTER SOURCE foo RENAME TO bar`... wait to return success.
2. HTTP POST, append to webhook source `foo` <- this is guaranteed to
return a 404.

### Motivation

Webhook sources are getting used more frequently now, so I wanted to
increase their supported throughput and thus stability.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - Improves the throughput of webhook sources.