more principled frontier checking to avoid flakes
This test assumed that the frontier jumps directly from {0} to {2} but
it is totally valid for the data to be committed in two steps, first
until frontier {1} and then until frontier {2}. The latter only happens
rarely due to the amount of data being very small and usually the
dataflows directly go to {2} but when it happened the tests would
produce false negative.

This PR fixes the problem by reading back the progress messages and
enforcing a more principled invariant checking by requiring that all
recorded frontiers beyond the expected upper are exactly one. This is
robust to additional steps being taken before the progress reaches the
expected frontier and additionally asserts that the frontier didn't move
*more* than it should have, something that the previous tests wouldn't
have caught.

Fixes #24063

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>