catalog: Add feature flag for catalog kind (#24029)
This commit adds a feature flag that determines what kind of catalog to
use. This feature flag is loosely modelled after the
"persist_txn_tables" feature flag. The final value of the catalog kind
flag is determined by looking at the following locations in order of
precedence:

  1. The persisted value of the "catalog_kind" system variable.
2. The default value for the "catalog_kind" system variable passed via
the CLI.
  3. The "catalog_store" CLI argument.

Like "persist_txn_tables", the "catalog_kind" feature flag is persisted
in both the "system_configuration" and the "config" collections. This
allows us to read the value before the catalog is fully opened.
However, correctly reading "catalog_kind" is even trickier. The value
is used to determine what catalog to use, but the value is stored in
the catalog. We take the following steps to overcome this:

  1. Start opening the catalog with the value from the CLI argument.
  2. Read the persisted "catalog_kind" flag.
  3. If the persisted flag or default value flag differs from the CLI
  argument, then TRY to switch the catalog implementation that we're
  using.

In practice this only works with catalog implementations that are
aware of both the persist and stash catalog implementations. So this
only works correctly with the `CatalogMigrator` implementation of the
catalog. This is fine because when we're ready to start migrating
environments, then we'll only use the `CatalogMigrator` implementation
and never the `Stash` or `Persist` implementations directly.

Works towards resolving #22392