clusterd: add command line flag for compute flow control
This commit adds a command line flag to control the `max_inflight_bytes`
number passed to `persist_source`s that feed compute dataflows.

For the production implementation this should be replaced with an LD
flag that's transmitted through `UpdateConfiguration` commands, but for
testing having a command line flag is way easier.