Testing for MFP fusion in Reduce
This commit adds more tests to validate that MFP fusion in Reduce is
able to handle complex MFPs in all reduction variants for which we can
trigger MFP fusion through SQL. Additionally, for each of these
variants, the tests also validate that errors are identified, reported,
and retracted properly. The commit opportunistically adds a bit more of
documentation to the fused MFP description as well.