persist-txn: do blocking read_at work in spawn_blocking
This is definitely blocking work and definitely should be run in
spawn_blocking. However, it seems to fix a bunch of spooky hangs and for
why that is, your guess is as good as mine.

While I'm in here, make the tokens on DataSubscribe more explictly a
`Vec<PressOnDropButton>`. This shouldn't affect anything, but is more
clear and matches what we recently started doing elsewhere in mz.