catalog: Always use catalog migrator implementation (#24087)
This commit updates the chosen catalog implementation to always be
`CatalogMigrator`. This means that every time we open the catalog we
will always either migrate the catalog contents to persist or rollback
the catalog contents to the stash. The default will be rolling the
catalog back to the stash, so the actual backing catalog store used will
remain as the stash.

As a consequence of this change, every time the stash catalog is opened,
we will also open the persist catalog. This will either initialize data
in the persist catalog with the set of default objects or migrate the
data in the persist catalog. No actual user catalog data will get stored
in the persist catalog. This will slightly slow down startup, but
opening the durable catalog is not the bottleneck of startup, so
hopefully it's not very noticeable. It will also allow us to validate
some very basic functionality of the persist catalog, without actually
relying on it for important data.

This change will additionally allow us to utilize the `catalog_kind`
system variable to migrate an environment to persist or rollback an
environment to the stash.

Resolves #22392