catalog: Fix migrate savepoint implementation
All existing environments will have an initialized stash catalog and an
uninitialized persist catalog. Therefore, reading a value before
opening the catalog will successfully identify the stash catalog as the
source of truth and read the value from the stash. However, when
opening a savepoint catalog, the `CatalogMigrator` will attempt to open
both the stash and persist in savepoint mode. The stash will succeed
but persist will fail because persist is uninitialized. This will
mistakenly cause the upgrade checker to think that an upgrade will fail
because the catalog is initialized (the result of a read from the
stash), but opening the catalog in savepoint mode fails.

This commit fixes this issue by allowing the `CatalogMigrator`
implementation to open a savepoint catalog when persist is
uninitialized and the migration direction is "rollback to stash". The
fix is to simply ignore the persist catalog and just open the stash
catalog in savepoint mode, since that is the target implementation.