catalog: Add escape hatch catalog kind flag
This commit adds a new value to the catalog kind flag called "legacy
stash". This variant uses the stash implementation directly without
first attempting to rollback from the persist implementation. This can
be used as an escape hatch in case the persist catalog is very broken
and unable to rollback successfully. Using this variant is potentially
very scary because it implies that we may be losing whatever data is
stored in persist. At that point we'd have to manually try and recover
the persist data and write it into the stash. Even though it's
potentially scary, if we ever do need this variant then we'll be glad
that we have it and are able to boot at all.