adapter: Create structure around plan optimization
Previously, plan optimization was done somewhere in the middle of
sequencing for plans that needed optimization. Each plan took care of
running optimization asynchronously if it needed to. This commit
creates an explicit stage in statement execution for plan optimization
after planning and before sequencing.

Having a separate stage will make it much easier to follow the
sequencing logic for each plan since it won't need to be broken down
into multiple pre and post optimization stages. Additionally, it
reduces the number of Coordinator messages, and it will allow all plans
to share the code for running optimization asynchronously. Currently,
the stage is run synchronously, but a later commit will run the stage
asynchronously.

This commit only updates the CreateView plan to take advantage of the
optimization stage as a PoC. Other plans can be updated in later
commits.