pgwire: fix perf regression with portal paging (#24093)
This commit fixes a performance regression introduced by #24093, which
allowed paging through portals in implicit transactions using the
extended query protocol.

To demonstrate the flaw in the original code, consider the following
sequence of extended query protocol messages, which repeatedly executes
a query named `q1`:

  1. Bind q1
  2. Execute q1
  3. Sync
  4. Bind q1
  5. Execute q1
  6. Sync

(1) starts an implicit transaction. (2) sets `txn_needs_commit = true`.
(3) commits the implicit transaction but crucially does *not* clear
`txn_needs_commit`. Processing of (4) sees that `txn_needs_commit` is
true and calls `StateMachine::commit_transaction`. This is a no-op,
because the transaction has already been committed and nothing new has
occurred, but it causes an additional roundtrip to the coordinator, and
thus the performance regression observed in #24149.

This commit adjusts the `StateMachine` to clear `txn_needs_commit`
whenever `end_transaction` is called, rather than just when
`ensure_transaction` handles `txn_needs_commit` being set. So now the
Sync message in (3) properly clears the `txn_needs_commit` flag, the
processing of the Bind message in (4) no longer needs an extra roundtrip
to the coordinator, and the performance regression disappears.

As an aside, I'm very impressed that the scalability benchmark was
sensitive enough to reliably report this regression. It makes sense that
only the benchmark of the simplest possible scenario (repeated `SELECT
1` commands) exposed the issue. More complicated queries will spend more
time in planning, optimization, and execution, and so the additional
coordinator roundtrip is negligible. But for `SELECT 1`, the additional
roundtrip is significant.

Fix #24149.