sql-parser: correct parsing of nested subqueries
This commit fixes the issue described in #24156, where an IN, ANY, ALL,
or SOME expression where the subquery on the right hand side was nested
within more than one pair of parentheses, e.g.:

    x IN ((SELECT 1))

was parsed incorrectly. Specifically, it was previously parsed
incorrectly as an Expr::InList(list: [Expr::Nested(Expr::Subquery)]),
whereas it was supposed to be parsed as an Expr::InSubquery.

Along the way, this commit fixes a number of places where the parser was
inconsistent about what tokens it allowed to start a query--most often
not recognizing `TABLE` or `SHOW` as valid starts to a query.

It also fixes a bug where expressions like:

    ((1) + 1, 1)

were incorrectly rejected by the parser--the notable feature being that
the first element of an implicit row constructor was an expression that
1) started with a nested expression and 2) followed the nested
expression with at least one other expression.

Fix #24156.