Specialize fused MFP in reductions
This commit specializes MFP fusion in reductions during MIR-to-LIR
lowering to identify when the MFP can be fully absorbed into the
Plan::Reduce, including the final projection. Identifying this special
case is useful as it is idiomatic in SQL to return aggregates that may
not include ones that are formed only to test conditions in the
HAVING clause. The analysis is now extremely conservative, and just
defaults to the general strategy whenever any mapped columns are
returned in the projection. This allows us to ensure that the special-
casing never regresses the memory requirements for the output
arrangement, not even by type changes in the output columns. When we
cannot match the special case, the general strategy that only absorbs
filters along with their supporting maps is used instead.