sql: ensure sources and sinks always refer to linked cluster
Because of the ordering in which we populate items in the
CatalogState, we need to ensure that sources and sinks always
explicitly refer to their linked clusters. Previously, it was possible
for items to elide a size specification, which is now impossible
to differentiate from an item that should be placed on the
default cluster. Unfortunately, we cannot just fix this up
after opening the catalog because it's possible that the default
cluster is incompatible with sources and sinks because it contains
more than one replica.