catalog: Update persist upper selection
This commit updates how the persist catalog chooses the next upper by
taking the max of the current time and the previous upper plus one.
Persist rate limits the consolidation of shards and uses the timestamps
of the uppers to determine how often to consolidate. The timestamps are
treated as epoch milliseconds. Therefore, if we only ever add one to
determine the next upper then consolidation will not run frequently
enough.