persist: remove maybe_heartbeat_reader call from Subscribe::next
This is vestigial from before we had the background heartbeat tasks.

The motivation is a very odd bug we're seeing in the persist catalog
impl. Writes to it are rare-ish, and so this Subscribe only gets new
batches rarely. This means the maybe_heartbeat call fires every time it
is run. Notably, this updates the last_heartbeat_ts on the read handle.
Then, the maybe_downgrade_since call inside the Listen (which would
actually update the since and seqno holds) gets skipped every time,
causing the shard since to get stuck indefinitely.