catalog: Update migrate savepoint implementation (#24181)
This commit is a followup to 2c9ef1534c4131c70f6ecb6085fae844d7dedec5,
which updates the migrate savepoint implementation to work when the
direction is "migrate to persist" and the stash is uninitialized. The
approach taken is identical to
2c9ef1534c4131c70f6ecb6085fae844d7dedec5, so see that commit message for
more details.

A scenario where we are migrating to persist and the stash is
uninitialized is extremely unlikely and probably impossible in
production. The `CatalogMigrator` implementation always opens the stash
catalog before the persist catalog and all prod environments have an
initialized stash catalog. Still, it's better to cover this scenario
than to ignore it.