catalog: Make persist debug QoL improvements (#24224)
This commit adds a couple of quality of life improvements to the
persist implementation of the catalog. Specifically it adds different
purpose strings to different uses of persist readers. Additionally, it
explicitly expires the old catalog implementations when migrating or
rolling back instead of waiting for them to expire from a timeout.