timely-util: separate range bounds from `Partitioned`
Previously the `Partitioned` timestamp type was opinionated about the
structure of the range of partitions, enforcing the existence of
positive and negative infinity elements (even if the inner type had a
perfectly fine minimum and maximum value) and the distinction between
singleton inclusive sets from range exclusive sets.

This PR simplifies the `Partitioned` timestamp to always encode an
inclusive range of elements made up of some type `P` that must have
extremas defined. The previous semantics of additional elements can then
be regained by using an appropriate `P` type and this is exactly what
the Kafka source timestamp has been changed to do.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>