ci: Bump Product Limits timeout
The test got slower with always_use_disk: https://github.com/MaterializeInc/materialize/pull/24239

This seems expected as we are now swapping out the lgalloc files.

An alternative would be to disable disk for this test, but I don't think
that's the proper way forward as we want to enable disk on production
too. Long-term we should rather evaluate which tests still make sense
and which will just be swapping to disk.

Seen in https://buildkite.com/materialize/nightlies/builds/5875