persist: allow CaA-ing serialized batches from the past
Before persist-txn, the only time we'd compare_and_append a batch
created somewhere else was in the multi-worker persist_sink, which might
mint some batches in each worker, exchange them into one place, and
compare_and_append all of them. This meant the version of code used to
generate the `ProtoBatch` was always the same as the one running the
compare_and_append.

Now, persist-txn commits batches to be appended to a WAL (something like
`(ShardId, ProtoBatch)>`), and then circles back later and appends
everything in the WAL to the physical shards. Usually this is done by
the same version of code, but it might not be if there's a crash in
between the two. This has now happened 11 times over the course of the
few weeks that persist-txn has been on in staging in "eager" mode, so it
happens but not constantly.

This change makes an internal `error!` check less restrictive to allow
batches from the past. The check was non-fatal, so this essentially just
turns the `error!` log into an `info!` log (which could be removed, but
seems nice to keep around given that this is low traffic). It does,
however, actually start failing if the batch was from the future, which
we probably should already have been doing.

I thought about making this check more restrictive, in particular only
allowing batches from the previous minor version, and in fact started
the branch out by implementing that. Persist-txn make sure to apply all
outstanding work from the WAL during startup and we require that all
environments get upgraded through each release, so this should work. But
then I remembered that persist already allows reading state from any
previous version, so that felt a little inconsistent.

Closes #23492